% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{afficher_tableau_cumuls}
\alias{afficher_tableau_cumuls}
\title{Affichage du tableau des cumuls de primes}
\usage{
afficher_tableau_cumuls(numero, libelles, e)
}
\arguments{
\item{numero}{Numéro du tableau}

\item{libelles}{Libellé des primes concernées ex. : PFR/IFTS}

\item{e}{Environnement retourné par \link{tableau_cumuls} ou NULL}
}
\description{
Affichage du tableau des cumuls de primes
}
\examples{
afficher_tableau_cumuls("5.1", "IAT/IFTS",
                        tableau_cumuls(test_prime(prime_IAT, prime_B = prime_IFTS, Paie_I, verbeux = FALSE)))
|Matricule |Annee |       Grade          |                  Régime                     |
|----------|------|----------------------|---------------------------------------------|
|010843    |2009  |ANIMATEUR TERRITORIAL |   IFTS 1 mois-IAT 10 mois-Cumul 1 mois      |
|010843    |2010  |ANIMATEUR TERRITORIAL |   IFTS 11 mois-IAT 0 mois-Cumul 1 mois      |
|010854    |2009  |REDACTEUR TERRITORIAL |   IFTS 9 mois-IAT 2 mois-Cumul 1 mois       |

}
