% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_EMD_plot.R
\name{build_EMD_plot}
\alias{build_EMD_plot}
\title{Title}
\usage{
build_EMD_plot(....)
}
\arguments{
\item{.x}{data.table}

\item{s}{numeric vector}

\item{t}{numeric vector}

\item{dt}{numeric}

\item{method}{string}

\item{boundary}{string}

\item{max.imf}{integer}

\item{noise.type}{string}

\item{noise.amp}{numeric}

\item{trials}{integer}

\item{stop.rule}{string}

\item{plot}{boolean}

\item{verbose}{boolean}
}
\value{
list
}
\description{
Title
}
