% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smpt_rmpp.R
\name{calcul_rmpp}
\alias{calcul_rmpp}
\title{RMPP
Calcul de la rémunération moyenne des personnes en place}
\usage{
calcul_rmpp(e)
}
\description{
RMPP
Calcul de la rémunération moyenne des personnes en place
}
