% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr4education.R
\name{cols4sum}
\alias{cols4sum}
\title{Funktion cols4sum}
\usage{
cols4sum(pData, pColname4Sum, pInCols)
}
\arguments{
\item{pData}{Ein Dataframe}

\item{pColname4Sum}{Der Name der neuen Spalte}

\item{pInCols}{Die Namen der Spalten, deren Summe berechnet werden soll}
}
\value{
Ein Dataframe mit der Summe der Spalten
}
\description{
Funktion cols4sum
}
\examples{

datensumme <- cols4sum(studentdata, "summe", c("aufgabe1", "aufgabe2", "aufgabe2"))

Spaltenbezeichnung in einem Vektor
colnames4data <-  c("aufgabe1", "aufgabe2", "aufgabe2")
datensumme <- cols4sum(studentdata, "summe", colnames4data)

}
