% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr4education.R
\name{copy4vec}
\alias{copy4vec}
\title{Funktion copy4vec(pVec)}
\usage{
copy4vec(pVec)
}
\arguments{
\item{pVec}{Ein Vektor}
}
\value{
Der kopierte Vektor mit gleichem Inhalten in den Komponenten
}
\description{
kopiert mit einer For-Schleife die Werte aus dem gegebenen Vektor pVec
in einen neuen Vektor und gibt diesen zurueck,
}
\examples{
copy4vec(c(1, 2, 3))

}
