% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr4education.R
\name{echo4vec}
\alias{echo4vec}
\title{Funktion echo4vec(pVarname,pVec)}
\usage{
echo4vec(pVarname, pVec)
}
\arguments{
\item{pVarname}{Der Name der Variablen als Zeichenkette}

\item{pVec}{Ein Vektor}
}
\value{
Zeichenkette, die mit Print ausgegeben wurde
}
\description{
Funktion echo4vec(pVarname,pVec)
}
