% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{Eliminer.duplications}
\alias{Eliminer.duplications}
\title{Eliminer les doublons en mémoire pour les calculs}
\usage{
Eliminer.duplications()
}
\value{
e   Environnement contenant les champs \code{Ndup}, nombre de bulletins doublonnés 
            et \code{Ndup2} (nombre de lignes doublonnées)
}
\description{
Eliminer les doublons en mémoire pour les calculs
}
