% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{Exporter.tables.sans.doublons}
\alias{Exporter.tables.sans.doublons}
\title{Eliminer les doublons dans les tables exportées sous Bases/Remunerations}
\usage{
Exporter.tables.sans.doublons()
}
\value{
e   Environnement contenant les champs \code{Ndup}, nombre de bulletins doublonnés 
            et \code{Ndup2} (nombre de lignes doublonnées)
}
\description{
Eliminer les doublons dans les tables exportées sous Bases/Remunerations
}
\note{
Les bases de doublons sont Paie.dup et Bulletins.paie.dup, les bases filtrées sont
      Bulletins.paie.uniq et Paie.uniq
}
