% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectifs.R
\name{filtrer.base}
\alias{filtrer.base}
\title{filtrer.base}
\usage{
filtrer.base(
  Base,
  grade,
  emploi,
  classe,
  service,
  libellés,
  agr,
  periode,
  statut,
  categorie,
  exclure.codes,
  quotite.nulle,
  type
)
}
\description{
filtrer.base
}
