% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliotheque.fonctions.paie.R
\name{filtrer_Paie}
\alias{filtrer_Paie}
\title{Filtrage d'une base de paye}
\usage{
filtrer_Paie(x, portee = NULL, Base = Paie, Var = "Code", indic = FALSE)
}
\description{
Filtre la base par une expression régulière sur libellés de paye ou par valeurs fixes sur une variable donnée.
Si le filtrage a une portee, l'ensemble des lignes de la portee (exemple "Mois") est conservé.
}
