% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr4education.R
\name{fuzzy2_implication}
\alias{fuzzy2_implication}
\title{Funktion fuzzy2_implication(pf1,pf2)}
\usage{
fuzzy2_implication(pf1, pf2)
}
\arguments{
\item{pf1}{Zahl oder Vektor mit Fuzzywerten}

\item{pf2}{Zahl oder Vektor mit Fuzzywerten}
}
\value{
Zahl oder Vektor mit Fuzzywerten
}
\description{
berechnet das fuzzylogische IMPLICATION mit der NICHT pf1 ODER pf2
von zwei Vektoren pf1 und pf2 von Fuzzywerte. Die Komponenten der Vektoren
sind reelle Zahlen zwischen 0 und 1. Analog zur Logik p->q <=> nicht p oder q
Grundlage f체r die fuzzy2_impcation() ist hier Negation der multiplikativen Verkn체pfung f체r UND
Dies liefert eine schw채chere Fuzzy-IMPLIKATION im Vergleich zu der fuzzy_implication() als Maximum
bzw. dem verwendeten fuzzy_or()
}
