% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr4education.R
\name{fuzzy_implication}
\alias{fuzzy_implication}
\title{Funktion fuzzy_implication(pf1,pf2)
berechnet das fuzzylogische IMPLICATION mit der NICHT pf1 ODER pf2
von zwei Vektoren pf1 und pf2 von Fuzzywerte. Die Komponenten der Vektoren
sind reelle Zahlen zwischen 0 und 1. Analog zur Logik p->q <=> nicht p oder q}
\usage{
fuzzy_implication(pf1, pf2)
}
\arguments{
\item{pf1}{Zahl oder Vektor mit Fuzzywerten}

\item{pf2}{Zahl oder Vektor mit Fuzzywerten}
}
\value{
Zahl oder Vektor mit Fuzzywerten
}
\description{
Funktion fuzzy_implication(pf1,pf2)
berechnet das fuzzylogische IMPLICATION mit der NICHT pf1 ODER pf2
von zwei Vektoren pf1 und pf2 von Fuzzywerte. Die Komponenten der Vektoren
sind reelle Zahlen zwischen 0 und 1. Analog zur Logik p->q <=> nicht p oder q
}
