% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr4education.R
\name{fuzzy_indicator}
\alias{fuzzy_indicator}
\title{Funktion fuzzy_indicator(pf1,pgrenze,pvalue0,pvalue1)}
\usage{
fuzzy_indicator(pf1, pgrenze, pvalue0 = 0, pvalue1 = 1)
}
\arguments{
\item{pf1}{Vektor mit Fuzzywerten zwischen 0 und 1}

\item{pgrenze}{die Grenze, ab der der Fuzzy-Indikator von 0 auf 1 gesetzt wird.}

\item{pvalue0}{der Ausgabewert, der bei dem Indikator 0 ausgegeben wird,}

\item{pvalue1}{der Ausgabewert, der bei dem Indikator 1 ausgegeben wird}
}
\value{
Ein Vektor
}
\description{
Funktion fuzzy_indicator(pf1,pgrenze,pvalue0,pvalue1)
}
\examples{
Vektor mit Fuzzywerten
fvec <- c(0.1, 0.9, 0.5, 0.7)

fuzzy_indicator(fvec, 0.5, 0, 1)
### Ausgabevektor (0,1,1,1)

fuzzy_indicator(fvec, 0.8, 0, 1)
### Ausgabevektor (0,1,0,0)

fuzzy_indicator(fvec, 0.5, "schlecht", "gut")
### Ausgabevektor ("schlecht","gut","gut","gut")

fuzzy_indicator(fvec, 0.8, "schlecht", "gut")
### Ausgabevektor ("schlecht","gut","schlecht","schlecht")

}
