% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr4education.R
\name{fuzzy_nand}
\alias{fuzzy_nand}
\title{Funktion fuzzy_nand(pf1,pf2)}
\usage{
fuzzy_nand(pf1, pf2)
}
\arguments{
\item{pf1}{Zahl oder Vektor mit Fuzzywerten}

\item{pf2}{Zahl oder Vektor mit Fuzzywerten}
}
\value{
Ein Vektor mit Fuzzywerten
}
\description{
berechnet die Negation des fuzzylogischen UND ("not and" bzw. "nand")
von zwei Vektoren pf1 und pf2 von Fuzzywerte. Die Komponenten der Vektoren
sind reelle Zahlen zwischen 0 und 1.
}
