% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generer_rapport.R
\name{generer_docx_odt}
\alias{generer_docx_odt}
\title{Générer les fichiers docx et odt à partir des fichier intermédiaires .tex générés pour le rapport pdf}
\usage{
generer_docx_odt(
  infile = "altair.tex",
  outfile = "altair.docx",
  outfile2 = "altair.odt",
  args = c("-V", "papersize=A4", "-V",
    "geometry:top=2cm,bottom=1.5cm,left=2cm,right=1.5cm", "-V", "urlcolor=cyan",
    "--highlight-style", "tango")
)
}
\arguments{
\item{infile}{Nom du ou des fichier(s) latex, par défaut altair.tex}

\item{outfile}{Nom du fichier docx de sortie, par défaut altair.docx}

\item{outfile2}{Nom du fichier odt de sortie, par défaut altair.odt}

\item{args}{Arguments à passer à pandoc comme options de forme}
}
\description{
Générer les fichiers docx et odt à partir des fichier intermédiaires .tex générés pour le rapport pdf
}
\note{
Permet de gagner du temps en recyclant des calculs déjà réalisés.
}
