% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generer_pdf.R
\name{generer_pdf}
\alias{generer_pdf}
\title{Générer le fichier pdf du rapport}
\usage{
generer_pdf(
  infile = "altair.tex",
  outfile = "altair.pdf",
  args = c("-V", "papersize=A4", "-V",
    "geometry:top=2cm,bottom=1.5cm,left=2cm,right=1.5cm", "-V", "urlcolor=cyan",
    "--highlight-style", "tango")
)
}
\arguments{
\item{infile}{le nom de sortie du fichier, par défaut altair.tex}

\item{outfile}{le nom de sortie du fichier, par défaut altair.pdf}

\item{args}{Arguments à passer à pandoc comme options de forme}
}
\value{
Retourne e chemin vers l'exécutable pandoc
}
\description{
Générer le fichier pdf du rapport
}
