% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{\%a\%}
\alias{\%a\%}
\title{Assignation dans l'environnement d'appel}
\usage{
x \%a\% y
}
\arguments{
\item{x}{Chaîne de caractères entre guillemets du nom de la variable assignée (gauche), comme pour \code{assign}}

\item{y}{Valeur. Variable assignante (droite)}
}
\description{
Assignation dans l'environnement d'appel
}
\examples{
"x" \%a\% median(1:100); cat(x)
}
