% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr4education.R
\name{hello}
\alias{hello}
\title{Funktion hello(name)}
\usage{
hello(pName)
}
\arguments{
\item{pName}{Name der Person, die den Text erhalten soll}
}
\description{
Funktion hello(name)
}
\examples{
hello("Hans")
}
