% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importer}
\alias{importer}
\title{Importer les données}
\usage{
importer(normaliser.colonnes = TRUE)
}
\arguments{
\item{normaliser.colonnes}{Enlever les accents des noms de colonnes}
}
\description{
Importer les données
}
\note{
Cette option vise à circonvenir un ancien bug de la bibliothèque \code{data.table}, désormais apparemment corrigé (2021)
}
