% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliotheque.fonctions.paie.R
\name{increment}
\alias{increment}
\title{Incrémente le numéro de tableau courant}
\usage{
increment()
}
\value{
Valeur incrémente de numero.tableau
}
\description{
Incrémente le numéro de tableau courant
}
