% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliotheque.fonctions.paie.R
\name{incrementer.chapitre}
\alias{incrementer.chapitre}
\title{Incrémente le numéro de chapitre courant}
\usage{
incrementer.chapitre()
}
\value{
Valeur incrementée de chapitre
}
\description{
Incrémente le numéro de chapitre courant
}
