% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliotheque.fonctions.paie.R
\name{inserer_script}
\alias{inserer_script}
\title{Insérer un script auxiliaire, indexé par une variable globale}
\usage{
inserer_script(
  chemin = NULL,
  index = 1,
  variable = "annee",
  gen = get("generer.rapport", envir = globalenv()),
  fonction = NULL,
  clean = TRUE,
  sync = get("sequentiel", envir = globalenv()),
  envir = NULL
)
}
\arguments{
\item{chemin}{Chemin du script R}

\item{index}{Vecteur numérique contenant les valeurs de la variable globale.}

\item{variable}{Vecteur de caractères contenant le nom de la variable globale dans le script auxiliaire.}

\item{gen}{Si \code{FALSE} alors se contente de sourcer le script auxiliaire selon \code{encodage.code.source}. Sinon intègre le rapport auxiliaire au format du rapport principal.}

\item{fonction}{Nom de fonction}

\item{clean}{Booléen. Nettoyer les fichiers intermédiaires}

\item{sync}{Caractère synchrone ou pas, géré par défaut par la variable globale #sequentiel}

\item{envir}{Environment de la function oi sinon NULL}

\item{seq}{Exécuter le script en mode sequentiel (si \code{TRUE}, resp. si \code{FALSE}, en mode parallèle)}
}
\value{
Valeur de la dernière variable globale \code{variable} instanciée. Effets de bord en sortie.
}
\description{
Insérer un script auxiliaire, indexé par une variable globale
}
