% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr4education.R
\name{komma2punkt}
\alias{komma2punkt}
\title{Funktion komma2punkt(pWert)}
\usage{
komma2punkt(pWert)
}
\arguments{
\item{pWert}{ist als Wert eine Zeichenkette, Vektor oder Spalte in einem Dataframe}
}
\value{
Die reelle Zahl die ggf. Komma durch eine R-lesbaren Dezimalpunkt
}
\description{
Diese Funktion "komma2punkt()" ersetzt ein deutsches Komma ","
in einer reellen Zahl (z.B. "3,5") durch einen Dezimalpunkt,
damit man mit den Werten in R rechnen kann (z.B. mit 3.5 ).
}
\examples{
Direkt eine Zeichenkette in die entsprechende Dezimalzahl umwandel
  komma2punkt("3,7")
gibt dann die reelle Zahl 3.7 zurück

Variable definieren und umwandeln
  wert <- "123,71"
  komma2punkt(wert)
gibt dann die reelle Zahl 123.71 zurück
Vektor definieren und umwandeln
  vec <- c("123,71","3,7",1.2, 51)
  komma2punkt(wert)
gibt dann den folgenden Vektor von reellen Zahlen c(123.71 , 3.7 , 1.2 , 51)  zurück
}
