% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr4education.R
\name{load_csv}
\alias{load_csv}
\title{Funktion load_csv(pFilename)}
\usage{
load_csv(pFilename)
}
\arguments{
\item{pFilename}{Der Dateipfad der CSV-Datei}
}
\value{
Ein Dataframe mit den Daten aus der CSV-Datei
}
\description{
Funktion load_csv(pFilename)
}
\examples{
df <- load_csv("meine_datei.csv")
studentdata <- load_csv("data/klausur.csv")

}
