% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr4education.R
\name{load_inout_csv}
\alias{load_inout_csv}
\title{Funktion load_inout_csv()}
\usage{
load_inout_csv(pFilename, pInCols, pOutCols)
}
\arguments{
\item{pFilename}{Dateipfad der CSV-Datei}

\item{pInCols}{Namen der Eingabespalten als Vektor mit Spaltenbezeichnungen}

\item{pOutCols}{Namen der Ausgabespalten als Vektor mit Spaltenbezeichnungen}
}
\value{
Ein Dataframe mit den ausgewählten Eingabespalten und Ausgabespalten
}
\description{
Diese Ladefunktion wird für die Auswahl von bestimmten Ein- und Ausgabespalten verwendet,
wobei nur die genannten Eingabe- und Ausgabespalten aus CSV-Datei extrahiert werden
Das die Datenspalten für numerische Berechnungen verwendet werden, werden
die Datenspalten ggf. in numerische Werte konvertiert.
ferner wird bei Rohdaten mit deutschem Komma in der Dezimalzahl "," durch "." ersetzt.
}
\examples{
incols <- c("Spalte1", "Spalte2","Spalte3")
outcols <- c("Spalte4", "Spalte5")

Die obigen Spaltenbezeichnung sollten als Spalten in der CSV-Datei existieren.
mydata <- load_inout_csv("klausur.csv", incols , outcols )

}
