% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliotheque.fonctions.paie.R
\name{merge_environment}
\alias{merge_environment}
\title{Fusionner un environnement à l'environnement global.}
\usage{
merge_environment(e)
}
\arguments{
\item{e}{Environnement}
}
\value{
Liste (non nommée) des objets ajoutés à .GlobalEnv
}
\description{
Fusionner un environnement à l'environnement global.
}
\examples{
c <- merge_environment(e)
}
