% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliotheque.fonctions.paie.R
\name{merge_environments}
\alias{merge_environments}
\title{Fusionner deux environnements.}
\usage{
merge_environments(e, envir)
}
\arguments{
\item{e}{Environnement 1}

\item{envir}{Environnement 2}
}
\value{
Environnement fusionné.
}
\description{
Fusionner deux environnements.
}
\examples{
c <- merge_environments(a, b)
}
