% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyramides.R
\name{produire_pyramides}
\alias{produire_pyramides}
\title{Produire les pyramides des âges par versant de la fonction publique}
\usage{
produire_pyramides(Filtre_bulletins, titre, versant = "", envir)
}
\arguments{
\item{Filtre_bulletins}{Fonction permettant de filtrer les bulletins sur les lignes de data.table}

\item{titre}{Titre de la pyramide.}

\item{versant}{Versant de la fonction publique ("FPT" ou "FPH")}

\item{envir}{Environnement de stockage des caractéristiques des âges (\code{nom.fichier.apres}, \code{nom.fichier.avant}, \code{res} quartiles de distribution des âges)}
}
\description{
Produire les pyramides des âges par versant de la fonction publique
}
