% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliotheque.fonctions.paie.R
\name{read.csv.skip}
\alias{read.csv.skip}
\title{Lecture d'une base CSV}
\usage{
read.csv.skip(
  x,
  classes = NA,
  drop = NULL,
  skip = 0,
  separateur.liste = separateur.liste.entree,
  separateur.decimal = separateur.decimal.entree
)
}
\arguments{
\item{classes}{Les classes ("character", "numeric") des variables en colonnes}

\item{drop}{Rang de la colonne à supprimer}

\item{skip}{Nombre de lignes à sauter en début de fichier (défaut aucune).}

\item{separateur.liste}{= separateur.liste.entree,}

\item{separateur.decimal}{= separateur.decimal.entree,}
}
\value{
Une base data.table
}
\description{
Lecture d'un fichier CSV et conversion en data.table.
Si sécuriser.types.sortie = TRUE, forçage des types en sortie.
}
\examples{
read.csv.skip(Base, separateur.decimal = ",")
}
