% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rendre.R
\name{rendre}
\alias{rendre}
\title{Générer le fichier latex du rapport, puis le pdf ou le docx et l'odt}
\usage{
rendre(
  fw = fig.width,
  fh = fig.height,
  d = dpi,
  keep = keep_md,
  clean = FALSE,
  to = "latex",
 
    from = "markdown+autolink_bare_uris+ascii_identifiers+tex_math_single_backslash-implicit_figures",
  args = c("-V", "papersize=A4"),
  texfile = "altair.tex",
  outfile = "altair.pdf",
  outfile2 = "altair.odt"
)
}
\arguments{
\item{fw}{Largeur des figures}

\item{fh}{Hauteur des figures}

\item{d}{Précision des graphiques en dpi}

\item{keep}{valeur booléenne : garder les fichiers intermédiaires md (par défaut la variable globale keep_md)}

\item{clean}{valeur booléenne : garder les autres fichiers intermédiaires (par défaut FALSE)}

\item{to}{Format de sortie (par défaut latex)}

\item{from}{Format d'entrée (par défaut une variété de markdown)}

\item{args}{Options à passer à pandoc}

\item{texfile}{Nom du fichier latex de sortie (altair.tex par défaut)}

\item{outfile}{Nom du premier fichier de sortie (pdf ou docx)}

\item{outfile2}{Nom du deuxième fichier de sortie optionnel (par défaut odt si précisé)}
}
\description{
Générer le fichier latex du rapport, puis le pdf ou le docx et l'odt
}
