% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliotheque.fonctions.paie.R
\name{Resume}
\alias{Resume}
\title{Statistiques descriptives de base (minimum, maximum et quartiles)}
\usage{
Resume(X, y, align = "r", extra = 0, type = "pond", id = "")
}
\arguments{
\item{X}{Vecteur de noms de variables en colonnes}

\item{y}{Les classes ("character", "numeric") des variables en colonnes}

\item{align}{Rang de la colonne à supprimer}

\item{extra}{Nombre de lignes à sauter en début de fichier (défaut aucune).}

\item{type}{Booléen (= FALSE).}

\item{id}{Identifiant de la table (pour le débogage).}
}
\value{
Une base data.table
}
\description{
Lecture d'un fichier CSV et conversion en data.table.
Si sécuriser.types.sortie = TRUE, forçage des types en sortie.
}
\examples{
read.csv.skip(Base, separateur.decimal = ",")
}
