% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliotheque.fonctions.paie.R
\name{Sauv.base}
\alias{Sauv.base}
\title{Sauvegarde d'une base}
\usage{
Sauv.base(
  dossier = "",
  nom = "",
  nom.sauv = nom,
  sep = separateur.liste.sortie,
  dec = separateur.decimal.sortie,
  environment = .GlobalEnv
)
}
\arguments{
\item{dossier}{Chemin du dossier dans lequel la base sera sauvegardée dans chemin.dossier.bases}

\item{nom}{Chaîne de caractères du nom de l'objet à sauvegarder}

\item{nom.sauv}{Chaine de caractères du nom du fichier .csv sans l'extension}

\item{sep}{(= separateur.liste.sortie)}

\item{dec}{(= separateur.decimal.sortie),}

\item{environment}{(= .GlobalEnv) environnement,}
}
\description{
Sauvegarde d'une base data.table sous forme de fichier CSV
Si sécuriser.types.sortie = TRUE, forçage des types en sortie.
}
\examples{
Sauv.base("Effectifs", "Base", "BaseDonnée")  -->  (dossier des bases) / Effectifs/BaseDonnée.csv
}
