% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr4education.R
\name{save_csv}
\alias{save_csv}
\title{Funktion save_csv(pFilename,pData)}
\usage{
save_csv(pFilename, pData)
}
\arguments{
\item{pFilename}{Der Dateipfad der CSV-Datei}

\item{pData}{der Dataframe mit Spaltenbezeichnung colnames}
}
\description{
Funktion save_csv(pFilename,pData)
}
\examples{
save_csv("meine_datei.csv", meine_data)
}
