% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showGithubREADME.R
\name{showGithubREADME}
\alias{showGithubREADME}
\title{Display GitHub README as iframe}
\usage{
showGithubREADME(repo, user = "averriK", height = "500px")
}
\arguments{
\item{repo}{Repository name (e.g., "gmsp", "kashima")}

\item{user}{GitHub username (default "averriK")}

\item{height}{Height of iframe (default "500px")}
}
\value{
Prints iframe HTML using cat() for results: asis
}
\description{
Display GitHub README as iframe
}
