% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smic.R
\name{smic.net}
\alias{smic.net}
\title{SMIC mensuel net x 3 utilisé pour la condition alternative sur les périodes non annexes}
\usage{
smic.net()
}
\description{
un salarié qui au cours de l'annee connait une rémunération supérieure à 3 fois le smic mensuel
est dans une période non annexe, quelle que soit la durée de cette période dans l'année.
remarque :  - on ne vérifie pas la continuité de cette période
            - on prend la moyenne annuelle des smic net déduction faite de la csg et de la crds, arrondie à la première décimale
}
