% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliotheque.fonctions.paie.R
\name{system.quote}
\alias{system.quote}
\title{Insérer des guillemets à chaque composante d'un vecteur de chaînes de caractères}
\usage{
system.quote(V)
}
\arguments{
\item{V}{Vecteur de chaînes de caractères.}
}
\value{
Le même vecteur avec des guillemets entourant chaque composante.
}
\description{
Insérer des guillemets à chaque composante d'un vecteur de chaînes de caractères
}
