% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliotheque.fonctions.paie.R
\name{Tableau.data}
\alias{Tableau.data}
\title{Tableau.data}
\usage{
Tableau.data(filename, colnames)
}
\arguments{
\item{filename}{Nom du fichier sous le répertoire data de la racine}
}
\description{
Tableau.data
}
