% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{tableau_cumuls}
\alias{tableau_cumuls}
\title{Tableau des cumuls de primes}
\usage{
tableau_cumuls(résultat, e = NULL)
}
\arguments{
\item{résultat}{Résultat retourné par la fonction \link{test_prime}}

\item{e}{Environnement, NULL par défaut}
}
\value{
Environnement avec champs \code{tableau} et \code{res}
}
\description{
Tableau des cumuls de primes
}
\examples{
tableau_cumuls(test_prime(prime_IAT, prime_B = prime_IFTS, Paie_I, verbeux = FALSE))


}
