% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliotheque.fonctions.paie.R
\name{tester.homogeneite.matricules}
\alias{tester.homogeneite.matricules}
\title{Teste si, dans une base, la proportion d'enregistrements Noms-Prenoms dont les matricules ne sont pas identiques
 reste inférieure à une marge de tolerance fixée (taux.tolerance.homonymie)}
\usage{
tester.homogeneite.matricules(Base)
}
\arguments{
\item{Base}{Base à tester}
}
\description{
Teste si, dans une base, la proportion d'enregistrements Noms-Prenoms dont les matricules ne sont pas identiques
 reste inférieure à une marge de tolerance fixée (taux.tolerance.homonymie)
}
