% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generer_rapport.R
\name{tex2pdf}
\alias{tex2pdf}
\title{Convertir le fichier tex du rapport en pdf}
\usage{
tex2pdf(
  infile = "altair.tex",
  outfile = "altair.pdf",
  args = c("-V", "papersize=A4", "-V",
    "geometry:top=2cm,bottom=1.5cm,left=2cm,right=1.5cm", "-V", "urlcolor=cyan",
    "--highlight-style", "tango"),
  keep = keep_md
)
}
\arguments{
\item{infile}{le nom de sortie du fichier, par défaut altair.tex}

\item{outfile}{le nom de sortie du fichier, par défaut altair.pdf}

\item{args}{Arguments à passer à pandoc comme options de forme}

\item{keep}{Garder les fichier .tex temporaires à la fin de l'exécution}
}
\description{
Convertir le fichier tex du rapport en pdf
}
