% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smpt_rmpp.R
\name{distribution_smpt}
\alias{distribution_smpt}
\title{distribution_smpt}
\usage{
distribution_smpt(Cat = NULL, Liste_statuts = NULL, S_ = NULL)
}
\arguments{
\item{Cat}{Catégorie des personnels (A, B, C)}

\item{Liste_statuts}{Statuts des personnels (ex.: c("STAGIAIRE", "TITULAIRE"))}

\item{S_}{Table temporaire calculée par #smpt}
}
\value{
Quartiles du SMPT
}
\description{
distribution_smpt
}
