% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generer_docx.R
\name{generer_docx_odt}
\alias{generer_docx_odt}
\title{Générer les fichiers docx et odt}
\usage{
generer_docx_odt(
  pandoc,
  infile = "altair.tex",
  outfile = "altair.docx",
  outfile2 = "altair.odt",
  args = c("-V", "papersize=A4", "-V",
    "geometry:top=2cm,bottom=1.5cm,left=2cm,right=1.5cm", "-V", "urlcolor=cyan",
    "--highlight-style", "tango")
)
}
\arguments{
\item{pandoc}{Chemin de pandoc
infile Nom du fichier latex, par défaut altair.tex
outfile Nom du fichier docx de sortie, par défaut altair.docx
outfile2 Nom du fichier odt de sortie, par défaut altair.odt
args Arguments à passer à pandoc comme options de forme}
}
\description{
Générer les fichiers docx et odt
}
