% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rendre.R
\name{rendre}
\alias{rendre}
\title{Générer le fichier latex du rapport, puis le pdf ou le docx et l'odt}
\usage{
rendre(
  fw = fig.width,
  fh = fig.height,
  d = dpi,
  keep = keep_md,
  clean = FALSE,
  to = "latex",
 
    from = "markdown-auto_identifiers+autolink_bare_uris+ascii_identifiers+tex_math_single_backslash-implicit_figures",
  args = c("-V", "papersize=A4", "-V",
    "geometry:top=2cm,bottom=1.5cm,left=2cm,right=1.5cm", "-V", "urlcolor=cyan",
    "--highlight-style", "tango"),
  filename = "altair",
  verbose = 0,
  sync = sequentiel
)
}
\arguments{
\item{fw}{Largeur des figures}

\item{fh}{Hauteur des figures}

\item{d}{Précision des graphiques en dpi}

\item{keep}{valeur booléenne : garder les fichiers intermédiaires md (par défaut la variable globale keep_md)}

\item{clean}{valeur booléenne : garder les autres fichiers intermédiaires (par défaut FALSE)}

\item{to}{Format de sortie (par défaut latex)}

\item{from}{Format d'entrée (par défaut une variété de markdown)}

\item{args}{Options à passer à pandoc}

\item{filename}{Racine du nom du fichier de sortie (altair par défaut)}

\item{verbose}{Par défaut, 0. Si fixé à 1, 2, 3, augmente la verbosité progressivement.}

\item{sync}{Booléen. Décrit le caractère synchone ou asynchrone de la génération des éléments de rapport. Géré par le système selon le paramètre global #sequentiel}
}
\value{
Retourne le vecteur des chemins des fichiers .tex temporaires
}
\description{
Générer le fichier latex du rapport, puis le pdf ou le docx et l'odt
}
