% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_Spectra.R
\name{smooth_Spectra}
\alias{smooth_Spectra}
\title{Title}
\usage{
smooth_Spectra(
  fs,
  A,
  method = "none",
  window = 5,
  po = 3,
  cv = FALSE,
  sigma = 2
)
}
\arguments{
\item{fs}{vector}

\item{A}{zp <- req(input$zp) |> as.double() vector}

\item{method}{string}

\item{window}{numeric}

\item{po}{numeric}

\item{cv}{boolean}

\item{sigma}{numeric}
}
\value{
list
}
\description{
Title
}
