% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smpt_rmpp.R
\name{smpt}
\alias{smpt}
\title{SMPT}
\usage{
smpt(Cat, Liste_statuts, type = "smpt net")
}
\arguments{
\item{Cat}{Catégorie des personnels (A, B, C)}

\item{Liste_statuts}{Statuts des personnels (ex.: c("STAGIAIRE", "TITULAIRE"))}

\item{type}{"smpt net", "smpt brut", "rmpp net", "rmpp brut"}
}
\value{
Tableau d'évolution du SMPT
}
\description{
SMPT
}
